% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNEX.R
\name{download_NEX}
\alias{download_NEX}
\title{NEX NC File download}
\usage{
download_NEX(variable, model, case, year)
}
\arguments{
\item{variable}{name ex. tas, tasmax, tasmin, pr}

\item{model}{name ex. ACCESS-ESM1-5, CESM2}

\item{case}{name ex. ssp245, ssp585}

\item{year}{target year after 2015}
}
\value{
Void
}
\description{
Download NEX satelliste image based on variable, model, case and year in working directory
}
\examples{
download_NEX("pr", "CESM2", "ssp245", 2025);
download_NEX("tas", "ACCESS-ESM1-5", "ssp245", 2028);
download_NEX(get_nex_variable_list()[1], get_nex_model_list()[1], get_nex_case_list()[1], 2028);
}
