% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNEXFilesList.R
\name{downloadNEXFilesList}
\alias{downloadNEXFilesList}
\title{Download metadata of NEX Server}
\usage{
downloadNEXFilesList()
}
\value{
VOID
}
\description{
Download NEX satelliste image based on variable, model, case and year in working directory
}
\examples{
downloadNEXFilesList();
}
