% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNEX.R
\name{extract_grid_NEX}
\alias{extract_grid_NEX}
\title{Extract GRID points provided Shape file}
\usage{
extract_grid_NEX(referenceShp, variable, model, case, year, fig = F)
}
\arguments{
\item{referenceShp}{Reference shape that want to extract}

\item{variable}{name ex. tas, tasmax, tasmin, pr}

\item{model}{Model name ex. ACCESS-ESM1-5, CESM2}

\item{case}{Case name ex. ssp245, ssp585}

\item{year}{target year after 2015}

\item{fig}{Figure as Boolean value as want to output the figure or not. Default value is FALSE}
}
\value{
Dataframe of all the points contains reference Shape Object in SP SpatialPointsDataFrame
}
\description{
Download NEX satelliste image based on variable, model, case and year in working directory
}
\examples{
data(bdDistrictShapes)
extract_grid_NEX(bdDistrictShapes, "tas", "ACCESS-ESM1-5", "ssp245", 2028);
extract_grid_NEX(bdDistrictShapes, get_nex_variable_list()[1], get_nex_model_list()[1], get_nex_case_list()[1], 2028);
}
