% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNEX.R
\name{extract_NEX}
\alias{extract_NEX}
\title{Extract NEX data for certain points}
\usage{
extract_NEX(targetPoints, variable, model, case, year)
}
\arguments{
\item{targetPoints}{Target points that want to extract in SpatialPointsDataFrame object}

\item{variable}{name ex. tas, tasmax, tasmin, pr}

\item{model}{Model name ex. ACCESS-ESM1-5, CESM2}

\item{case}{Case name ex. ssp245, ssp585}

\item{year}{target year after 2015}
}
\value{
Void
}
\description{
Download NEX satelliste image based on variable, model, case and year in working directory
}
\examples{
data(bdDistrictCentroids)
extract_NEX(bdDistrictCentroids, "tas", "ACCESS-ESM1-5", "ssp245", 2028);
extract_NEX(bdDistrictCentroids, get_nex_variable_list()[1], get_nex_model_list()[1], get_nex_case_list()[1], 2028);
}
