% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNEX.R
\name{get_nex_case_list}
\alias{get_nex_case_list}
\title{NEX Case list}
\usage{
get_nex_case_list()
}
\value{
list of array with model names
}
\description{
Get the list of all available cases to download
}
\examples{
get_nex_case_list();
}
