% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNEX.R
\name{get_nex_model_list}
\alias{get_nex_model_list}
\title{NEX models list}
\usage{
get_nex_model_list()
}
\value{
list of array with model names
}
\description{
Get the list of all available models to download
}
\examples{
get_nex_model_list();
}
