% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNEX.R
\name{get_nex_variable_list}
\alias{get_nex_variable_list}
\title{NEX variable list}
\usage{
get_nex_variable_list()
}
\value{
list of array with variable names
}
\description{
Get the list of all available variables to download
}
\examples{
get_nex_variable_list();
}
