% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNEX.R
\name{visualize_NEX}
\alias{visualize_NEX}
\title{Visualize NEX data for certain shape files}
\usage{
visualize_NEX(referenceShp, variable, model, case, year, fun = mean)
}
\arguments{
\item{referenceShp}{Reference shape file object that want to visualize}

\item{variable}{name ex. tas, tasmax, tasmin, pr}

\item{model}{Model name ex. ACCESS-ESM1-5, CESM2}

\item{case}{Case name ex. ssp245, ssp585}

\item{year}{target year after 2015}

\item{fun}{Target Function that you want to aggregate. Default mean}
}
\value{
Void
}
\description{
Download NEX satelliste image based on variable, model, case and year in working directory
}
\examples{
data(bdDistrictShapes)
visualize_NEX(bdDistrictShapes, "tas", "ACCESS-ESM1-5", "ssp245", 2028);
visualize_NEX(bdDistrictShapes, get_nex_variable_list()[1], get_nex_model_list()[1], get_nex_case_list()[1], 2028);
}
